Initial draft code
This is the initial core python script that reads a csv file exported from Excel (exported from the Varioskan microplate reader). It generates a Microplate Experiment object that can do basic file reading, plotting, and (importantly) reformat the data to a csv file pivot table with variable wavelength as the first column and columns for each sample there is data for in the next columns (thus easy to graph in Excel etc.). Currently draft but seems to work for photometric, emission, and excitation data.